requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2
python-multipart==0.0.6
httpx==0.25.2
geopy==2.4.1
scikit-learn==1.3.2
openai==1.3.7
python-dotenv==1.0.0
ortools==9.7.2996